requirements.txt
accelerate==0.27.0.dev0
appdirs==1.4.4
backcall==0.2.0
biopython==1.81
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
daal4py==2023.1.1
decorator==5.0.7
DSSPparser==0.1.3
einops==0.6.1
et-xmlfile==1.1.0
fair-esm==2.0.0
filelock==3.9.0
flit_core==3.6.0
fonttools==4.25.0
fsspec==2023.12.2
gmpy2==2.1.2
huggingface-hub==0.20.2
idna==3.4
importlib-resources==5.2.0
IProgress==0.4
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.1.2
joblib==1.1.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.4.4
lmdb==1.4.0
MarkupSafe==2.1.1
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
munkres==1.1.4
networkx==3.0
ninja==1.11.1
numpy==1.23.5
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
ply==3.11
pooch==1.4.0
prompt-toolkit==3.0.18
psutil==5.9.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.8.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==22.0.3
rdkit-pypi==2022.9.5
regex==2023.3.23
requests==2.28.1
safetensors==0.4.1
scikit-learn==1.2.2
scikit-learn-intelex==20230426.111118
scipy==1.10.1
setuptools==65.6.3
sip==6.6.2
six==1.16.0
sympy==1.11.1
threadpoolctl==2.2.0
tokenizers==0.15.0
toml==0.10.2
torch==2.0.0
torch-cluster==1.6.1
torch-geometric==2.3.0
torch-scatter==2.1.1+pt20cpu
torch-spline-conv==1.2.2+pt20cpu
torchaudio==2.0.0
torchvision==0.15.0
tornado==6.1
tqdm==4.65.0
traitlets==5.0.5
transformers==4.37.0.dev0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.5
wheel==0.38.4
xmltodict==0.13.0
zipp==3.11.0
